requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
httpx==0.27.2
Pillow==10.4.0
numpy==1.26.4
openai>=1.40.0
